adds unauth + protocol reqs
