+++ Sources/Authentication/Utilities/AuthenticationError.swift
