+++ Sources/Authentication/Utilities/StringConvertible.swift
